Allow turning off hoerbert.xml regeneration
If the user will stick to version 2.0 and never go back to 1.0, there is no point in regenerating hoerbert.xml.
The only thing is: Either there has to be hoerbert.xml AND hoerbert.bak on the card, OR no hoerbert.xml at all.
If there is a hoerbert.xml and no hoerbert.bak, then migration will start.